refactor(tests): refactor tests for data validators and pytest config
